FEAT nav links on toolbar
The toolbar at the top now also features navigation buttons to the list
of tokens and to the enroll token component, in addition to the logout
button.

Since enrolling is on the same level as listing an managing existing
tokens, the route is now /enroll and not /tokens/enroll.

Refers to LINOTP-673

Change-Id: I07f9d1c8b9c10325efbd2442aff960e598aab654