FEAT display token list as material accordeon
The token details and available actions are now shown when the user clicks on the desired
token and the panel expands.

This means that the functionality previously available in TokenComponent
(deleteToken and showing Token data) have been moved to
TokenListComponent.

The routes have also been updated: there is no more tokens/:id.
However, tokens/:id/activate still exists. This situation will have to
be verified once we integrate the token activation functionality.

Refers to LINOTP-684

Change-Id: I692ab29902daf7d9e9518b3336161ef2f0039b6e