FEAT reorganize enrollment steps
The stepper is now vertical instead of horizontal. This allows us to
have an unlimited amount of steps.

Each of the steps should only address one
issue so there are more steps now.

The step for setting a pin was made optional.

Once we implement the functionality for testing the new token, we can
display a success message. For now the test step displays the OTP input
form and a hint that, should something go wrong with the test, the user
can return to step 2 and try out another app, or contact her
administrator.

Refers to LINOTP-673

Change-Id: I0f3b8a4ca1f6b0682e4028ab208e3f07dab8f4f7