feat: introduce password token type
The Password Token is now recognized by the frontend, as well as the
respective enrollment permissions.

In the unit test that verifies the mapping of policies to permissions we
now check not only the delete policy but also the password enrollment
policy and all other policies known to the frontend.