feat: enable/disable token via card menu
The user can enable and disable tokens via buttons in the card menu,
provided they have the required permissions.

If the user does not have permissions to enable or disable a token, the
button should not disappear from the menu. Instead, it is disabled and
displays a message about the lack of permissions on mouseover.