feat: let token card call individual dialogs
Instead of chaining the dialogs by calling the test OTP dialog from the
enrollment dialog, we let the component handle the results of the first
dialog and then call the second one.

This prevents one dialog that has finished all its
steps from remaining in the background while the second dialog is
active.