chore: use debian based docker images for CI runs
With the previous mode of having a mixture of Debian and Alpine based
images sharing the cached `node_modules` would not always work.
Particularly the case where we are sharing pre-compiled binaries for one
distribution and execute them on the other would fail without yarn
noticing or reinstalling those. The easiest solution is to have all CI
stages share the same base image.

a bit related to LINOTP-686