chore: add translation infrastructure
a package.json script `extract-i18n` is introduced to extract translations
from templates and and typescript files.  The `postextract-i18n` script
is executed afterwards automatically. It merges the new translations
into the existing translated language files.

We use the i18n-polyfill package that mimics the speculative new api
that will be available later with angular to translate strings in
typescript files. The I18n service is available in the application.

Change-Id: Ia21e6ecdb256a729a394f02a4b4e20fe0d75afcb