Revert "Merge branch 'feature/webprovision-policy' into 'master'"
This reverts commit 436011990e0a95107e8a7bc65e9a952f5572d56b, reversing
changes made to cf278eded00eef7b54c9f154f37ccd7c39944817.

We are currently enrolling HOTP/TOTP tokens with a configuration
compatible to the webprovisionGOOGLE enrollment. However, we are still
using the /enroll/ endpoint, whose usage permissions are defined by the
enrollHOTP and enrollTOTP policies.

The reverted merge request only changed the policies/permissions, but
not the enrollment endpoint.