fix: OATH enrollment type differs from token type
When enrolling OATH tokens over the /webprovision endpoint, the supplied
token type is "googleauthenticator" or "googleauthenticator_time",
however, the resulting token type is "hmac" or "totp" instead.

Because we were assuming the resulting token type was the same as the
enrollment type, the retrieved token list was not displaying the OATH
tokens appropriately, as their types were not being recognized. The
solution was to set a distinct "enrollmentType" for this special case.

Co-authored-by: Andreas Rammhold <andreas.rammhold@keyidentity.com>