fix: typescript compile target to new for qrcode lib
Angular defaults to the newer compile target `es2015`. The dependency
angularx-qrcode is currently not supporting this compile target.
Therefore we revert to target `es5`. Development regarding this issue is tracked here:
https://github.com/Cordobo/angularx-qrcode/issues/53

Once v2 is released we must switch back to es2015 builds to ensure
latest and greates optimizations can be used by angular.