refactor: move choice of MFA login token to login()
Instead of receiving the serial of the first available token, the login
method now gets a list of all available tokens for the user. This will
allow us later to let the user choose the token from the list. For now,
the login method still chooses the first token automatically as an
intermediate step.