refactor: split auth service into login and session services
We currently have a cyclic dependency between the auth service and the
system service. It's not nice, but the application is still able to
instantiate and inject the services.

In the upcoming features we will introduce another dependency, as the login
logic will require access to the token service. Since that would
definitely break the dependency injection (we tested it!), we decided to
split the AuthService in two: LoginService - responsible for making
backend requests related to login and logout - and SessionService -
which handles the login state and user permission updates.

In this commit the session service is still called AuthService. For
better readability of the changes we rename it in the next commit.