feature: support testing of transaction-based, non chall.-resp. tokens
OTP-generating tokens that first need to trigger a transaction to send
the OTP to the user, such as SMS and Email tokens, receive a
TransactionDetail object in the 1st response to the backend `/verify`
request.

The TestService now recognizes the TransactionDetail and can pass it
along to any component starting a token test for said token types.