feature: test SMS and Email tokens in Test OTP Dialog
The TestOTPDialogComponent triggers a test transaction on
initialization. This stores a transaction details object.  When the user
later submits the OTP, a transaction ID is sent along if there was one
provided by the backend in the transaction details, as is the case with
SMS and e-mail tokens.

When the user submits the OTP, if the token is in transaction mode it
also sends the transaction ID.