feature: close all dialogs on automatic logout
When the user session times out or is otherwise invalid, we do an automatic logout procedure, that moves the user back to the login so that he can start again. If this would happen during a dialog is opened, it would stay open above the login screen.
This is hereby changed and can nolonger happen.

Refers to LINSLEF-55