feature: create an mOTP token
The user can start the process to generate an mOTP token by clicking on
the 'create mOTP' button on the enrollment grid, provided the
appropriate permission has been set.

Clicking on said button opens a dialog that guides the user through the
required steps: entering a password (seed) and an mOTP pin. Optionally a
description and an additional security pin can be set. At the end of the
process, the user may test the token if they have the required
permisisons.