fix: token activation under pin=password policy
we now use /validate/check/ with param serial instead of /validate/check_s because of incompatibliity problems. Now it should be possible to activate the token using the user password as pin if so configured in linotp. This was previously not possible because check_s refuses to fight with password pin configuration.

Closes LINSELF-57