feature: recognize edit_<sms|email> policy during enrollment
edit_sms and edit_email policies configure if the user is allowed to change the address/number during the token enrollment process. Instead of the input we show the user only a paragraph with the address/number that will be ued.

It is required to send the target address/number in the API request even if the user is not allowed to chose a different value because LinOTP expects the users default to be sent without a modfication. The key can not be omitted.

Closes LINSELF-12
Closes LINSELF-13