refactor: extract mapping of token object
The functionality of mapping the backend object to a frontend object
will be needed in more than one locations. We extract it from the
mapTokenResponse function which converts an entire response object
containing a list of tokens. It will soon be used by the login request
to map the list of second factors :)