fix: do not tell token will be deleted when it won't
When a user does not have permission to delete a token, but cancels the
enrollment of a token at a point of the setup process when the token has
already been created on the backend, the confirmation message was saying
that the token would be deleted.

We now check for the `delete` permission, and if it is not set, the user
is instead informed that the token will not be in a usable state.