test: remove usage of deprected TestBed.get using a new testing helper method
TestBed.get was deprecated for a new variant TestBed.inject that does
the same thing but is now properly typed. This brought up some problems
where we correctly typed the result to be a jasmine.SpyObj<Service> type
if we retrieved a mocked spy with TestBed.get.

A helper method called getInjectedStub() was added to help with the
typing. It is now retrieved type safe. Please make sure that you only
use it for a service that is mocked. We typically do this with the
spyOnClass helper method. If you would this new method on a non-mocked
object, the resulting type would suggest that you can use the mock
methods which will fail.