feature: display warning on IE and legacy Edge
In order to display the warning we have to replace the Angular app with
plain HTML and CSS, as the older browsers do not support modern Angular by
default, with standard settings.

This means we need a script for each translation language to replace the
inner HTML, and therefore the build configuration had to be adapted for
this effect.