feature: make SystemInfo reusable without separate API requests
This allows to use the system info on different components other than
the login component without creating additional calls to the backend.
Therefore an observable is created in the constructor that is executed
on the first subscription and only replays the data afterwards. It is
first replying with a localStorage cache version and replaced later with
the live data from the backend.