fix: stacking redirect param that prevents redirect on successful login
The AuthInterceptor service uses the LoginService.handleLogut method to
handle the app logic if a route is targeted but a valid session is
missing during the navigation. Previously the handleLogout method would
only look at the current route url and therefore not correctly detect
the redirect target on navigation.

A unit test is not really possible without exaclty reengineering the
same arguments that the handleLogout method is accessing. This would
not prove anything is fixed... A full app setup would be required but
we currently do not have e2e tests.

Closes LINSELF-75