fix: dynamicly generate enrollment permissions list for the token list
This way, the enrollment grid will always be shown if any token
enrollment policy is set because we no longer rely on a manually
maintained list that will probably go out of sync if a new token
type is added.

This also fixes a problem where the enrollment grid was shown
without any enrollment policies being active because the permission
`verify` was mistakenly listed as an enrollment policy.