feature: reword set PIN according to context
Token card menu entry should say "change PIN", while the dialog is still
titled "Set a new PIN". In the enrollment dialog it says "Set PIN" in
the button, because we are certain that there was no PIN set by the user
before.

It's also now "die PIN" and not "der PIN" everywhere in the German
translation.