feature: operations service shows notifications on failure
To make behavior across the app more consistent, the operations service
now displays a error toast if the operation failed, like e.g. the
enrollment service does. This allows us to eventually handle all sorts
of backend errors in a unified way on the service side.

This also means that the entities calling the operations only need to
handle the success case, so we updated the token card component
accordingly.