feature: services handle http errors and failures equally
For all purposes, to the user a server error and an unsuccessful
operation should result in the same behavior: a notification of
unsuccessful operation. We now make sure all services catch any sort
of negative response from the server and display the notification
themselves, so the components do not have to care about what exactly
went wrong.