fix: update token list after activating a token
We now enforce the token update to happen when the observable is fully
processed.

Curiously the tests were not catching this issue. The order of the token
update seemed to be wrong before, so that there were indeed two token
updates for the push and qr token, but the last one happened too early.