fix: side margin in app component not working
margins do not cooperate well with flexbox designs. Using border-box with
padding achieves the desired effect.

Because the box-sizing steals some width from the elements, we increase
the max-width approximately with the lost width. We do not exaclty
measure the lost width because the 800px width are only a coarse design
decision initially.

Some smaller fixes to the toolbar ensure that the toolbar is drawn around
the content even if the content is bigger than 100%

Closes LINSELF-106