fix: permission checks not updating on permission change
Permission checked in a controller and not via the *ngxPermissionsXXXX
directive are not automatically updating if permissions are reloaded. The
directive does this automatically. With the new feature of refreshing
permissions, permission states would not be consistent because the
controller loaded permissions are not updating.

A new login service method `hasPermission$` is herby introduced to check
a given permission over time. The subscriptions need to be unsubscribed
on component destroy (implement OnDestroy interface) because the
hasPermission$ subscriptions are not self closing.

Problem found and fixed during development of LINSELF-98 where we need
to check for another permission that needs to be up to date as well.