fix: history search is no longer implicitly submitted
Changing the sort order or pagination previously would have submitted
the currently entered search term. The user now has to explicitly submit
the search.

To show the user whether the current search term is submitted or
pending, the serach button is only shown if the search is not submitted.

The clear search button now has a translated tooltip and is shown if the
currently shown history view is filtered by the query.

This is the last part of and therefore closes LINSELF-105