fix: login component stuck at step one in mfa_login
For users that have exactly one token, the UI is supposed to directly
jump to the challenge-answer step by selecting the token.
This request would not be done because the method chooseSecondFactor
thinks that it was called mistakenly as a duplicate. We now clear the
awaitingResponse flag first thing in the api callback to fix this problem.

Closes LINSELF-119