chore: use default link resolution for routes
Relative paths like `./tokens` have to be changed to e.g. `../tokens` to
work with the new default `relativeLinkResolution: 'corrected'`. Setting
the value to `legacy` fixes the breakage. This was automatically done by
the update script when upgrading to v11.

However we do not have such paths, and so it's safe to use the default.

More info:
https://v11.angular.io/api/router/ExtraOptions#relativeLinkResolution