feature: deep link to token enrollment
A newly created enrollment component decides which dialog to open
depending on the token type. Said component is reached via the route
`/tokens/enroll/:type`.

This means that the enrollment grid is no longer responsible for that
logic, and the buttons in the enrollment cards simply navigate to the
respective route.