fix: pass token type to test and activate dialogs
Through some previous refactoring we missed that the data type
EnrolledToken was being abused as the parameter for the Test and
Activation dialogs. The data structure matched when triggering the test
from the enrollment dialog, but not from the token card dialog, where we
have a SelfserviceToken with a slightly different structure.

We fix this issue and simplify the communication by passing only the
values relevant to the testing and activation, namely the serial and the
token type.

The "activate" boolean was also not used anymore and was removed. It
comes from the time where the test and the activation both took place in
the same dialog.