fix: reload token list on enrollment state change
Push and QR tokens should show up differently on the token list if they
have been created but not paired, or paired but not activated.

Interrupting the enrollment or activation process would in almost all
cases lead to a wrong display of the token list, so moving the reload to
the individual enrollment dialogs will let us fix it.

In order to do this, we must move the update trigger into the enrollment
dialogs of all tokens, instead of having it only done by the component
that opens the dialog.