chore: replace relative imports with path aliases
Project-internal imports relative to the file are now replaced with the
path aliases defined in the tsconfig.json file.

The only exception is the case where imports are in the same directory.
Usually these files belong together (e.g. tests for a component or
service importing that component or service), and anything other than
"./" cannot be any shorter nor more useful.