chore: Add `root` directive to nginx configuration to fix reloading issue.
When a “deep link” reload is triggered (e.g. by hitting F5 in popular
web browsers), nginx cannot serve the resource in question directly
because no file on the server corresponds to it. Instead, the
operation is punted to the SPA via the `LANGUAGE/index.html` file (see
the `try_files` directive in the nginx configuration).

`try_files`, however, looks for the files in question in the document
root (it doesn't care about `alias` directives). There is no `root`
directive in the current nginx configuration at all, so nginx falls
back to `/etc/nginx/html`, which is obviously bovine excrement.

The fix is to add a proper `root` definition.