fix: Make nginx emit relative URIs in `Location:` headers.
By default, if an HTTP request leads to a redirection response and the
new URI in the `Location:` header starts with a slash, nginx prepends
the request's URI scheme and “netloc” (server name and port number) to
that URI. This can lead to issues if the selfservice is behind a
reverse proxy and the rewritten URI in the `Location:` header
describes a resource that doesn't exist or can't be reached by the
client's browser.

We would like relative URIs to stay relative so the browser will use
the original protocol scheme and netloc to do the redirection. This
decreases the likelihood of errors and simplifies reverse-proxy
configuration.