fix: Fix redirection directive for selfservice base URI.
If the redirected URI ends with a slash, then the URI being redirected
to should generally also end with a slash. This will save us one round
of redirections where the final redirection just sends us from
something like `/selfservice/en` to something like
`/selfservice/en/`. (This is frankly a bit of a rookie mistake.)

While we're at it, we also make the redirection temporary rather than
permanent. Permanent redirections are for resources that move from A
to B; in the case of a language preference, a temporary redirection
makes more sense.