fix: Change Apache restart operation in postinst script.
Simply saying “systemctl restart apache2” will fail if you install the
package in a place that doesn't have systemd (e.g., in a Docker
container). We change the script to invoke `systemctl` only if we're
actually running under systemd (as indicated by the presence of the
`/run/systemd/system` directory). We also use `deb-systemd-invoke`
rather than `systemctl`, which is what we're supposed to do in a
postinst script, mostly so the Right Thing will happen if the service
was disabled. This doesn't really apply to us, and we might get away
with calling `systemctl reload-or-restart` after all as long as we
have the directory guard in place, but for now we're doing things by
the book.