feature: Enable asset overrides in nginx for the Docker image.
If local administrators want to override static assets for the
selfservice (such as CSS style sheets or logos), they can put their
custom versions in a directory and mount that on `/custom-assets` in
the container.

When looking for a static asset with an URL like
`…/selfservice/de/assets/logo.png`, nginx does the following:

1. If the file `/custom-assets/de/logo.png` exists, it is returned.
2. If not, and if the file `/custom-assets/logo.png` exists, that file
   is returned.
3. If that isn't the case, and the file
   `/usr/share/nginx/html/en/assets/logo.png` exists, that file is
   returned.
4. An HTTP 404 error is returned.

Note that `/custom-assets` is a symlink to
`/usr/share/nginx/html/custom-assets`. This is so we can use
`try_files` with a common prefix set up using `alias`. (You need an
`alias` or `root` definition in scope for nginx to be able to work. It
would be a security issue if nginx could return arbitrary files on the
system based on absolute path names.) Even so, if you mount a volume
on `/custom-assets`, its content will show up in
`/usr/share/nginx/html/custom-assets`, so that is OK.