chore: depend on apache2 in Debian package to fix postinst error
Because we do not depend on apache2, it is not yet completely configured when installed as a recommended package when our postinst runs. It is however partially installed, which makes our postinst fail.

The easy fix is to just strictly depend on it. We do not really plan to distribute the selfservice as a Debian package without using it together with linotp and apache2 anyway.

Closes LINSELF-133